requirements.txt
pika==0.9.14
requests==2.7.0
python_logstash>=0.4.5
retrying==1.3.3